Modified when the figureLogController was called from canvasController
Moved the ListViewCell
Updated width of figure log
