+++ src/main/resources/FigureLog.fxml
@@ -6,11 +6,11 @@
 <?import javafx.scene.control.ListView?>
 <?import java.util.ArrayList?>
 
-<AnchorPane fx:id="root" prefWidth="300.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1"  fx:controller="apaintus.controllers.FigureLogController">
+<AnchorPane fx:id="root" prefWidth="225.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="apaintus.controllers.FigureLogController">
 <children>
 	<VBox alignment="CENTER">
 		<Label text="Drawn Shapes History"/>
-		<ListView fx:id="figureList" prefWidth="300.0" editable="true"/>
+		<ListView fx:id="figureList" prefWidth="225.0" editable="true"/>
 	</VBox>
 </children>
 </AnchorPane>
